requirements.txt
streamlit==1.54.0
ollama==0.6.0
pymupdf==1.27.1
transformers==4.57.1
faiss-cpu==1.12.0
pillow==11.3.0
numpy==2.2.6
scipy==1.16.2
openai-whisper==20250625